+++ NEWS.md
@@ -3,3 +3,9 @@
 * Initial version of the package based on `get_colormind_colors()` and `get_colormind_models()`.
 * Added a `NEWS.md` file to track changes to the package.
 * Made a logo and a github card
+
+# colormind 0.0.0.9001
+
+* made get functions polite (choked rate to 1/3 req/sec, provided default user-agent)
+* added some tests for `get_colormind_models()`
+* added dependency on `curl` for the purpose of checking for internet.